fix: older parser versions were linking only inside frame
