+++ app/views/versions/_list.html.erb
@@ -31,13 +31,13 @@
 
   <% versionable.versions.limit(10).desc(:created_at).each_with_index do |version, index| %>
     <div class="version <%= 'active' if @version.try(:id) == version.id %>">
-      <%= link_to version.message || 'No message',  send(:"#{versionable_path}_path", versionable, version) %>
+      <%= link_to version.message || 'No message', send(:"#{versionable_path}_path", versionable, version) %>
       <p><%= localize_date_time(version.created_at) %> by <%= version.user.try(:first_name) %></p>
       <%= environment_tags(version, versionable) %>
     </div>
   <% end %>
 
-  <%= turbo_frame_tag :more_versions, src: async_load_path do %>
+  <%= turbo_frame_tag :more_versions, src: async_load_path, target: :_top do %>
     <%= image_tag image_path('sj-spinner.gif', width: 40, height: 40) %>
     <strong>Loading more versions..</strong>
   <% end %>
